updating dev

+++ src/execute/here_doc_cleanup.c
@@ -3,15 +3,22 @@
 void cleanup_heredoc_files(t_heredoc_data *heredocs, int count)
 {
     int i;
-
-    i = 0;
+	
+	i = 0;
+    if (!heredocs)
+        return;
     while (i < count)
     {
-        unlink(heredocs[i].temp_file);
-        free(heredocs[i].temp_file);
+        if (heredocs[i].temp_file)
+        {
+            unlink(heredocs[i].temp_file);
+            free(heredocs[i].temp_file);
+            heredocs[i].temp_file = NULL;
+        }
         i++;
     }
     free(heredocs);
+    heredocs = NULL;
 }
 
 int free_initialized_heredocs(t_heredoc_data *heredocs, int count)
